Moved pages to the dapp domain for the frontend-counter-react package.
