+++ packages/frontend-counter-react/src/components/App.tsx
@@ -7,8 +7,8 @@ import { BrowserRouter, Route, Routes } from 'react-router'
 import { APP_NAME } from '~~/config/main'
 import { getThemeSettings } from '~~/helpers/theme'
 import useNetworkConfig from '~~/hooks/useNetworkConfig'
-import CounterPage from '~~/pages/CounterPage'
-import IndexPage from '~~/pages/IndexPage'
+import CounterPage from '~~/dapp/pages/CounterPage'
+import IndexPage from '~~/dapp/pages/IndexPage'
 import ThemeProvider from '~~/providers/ThemeProvider'
 import '~~/styles/index.css'
 import { ENetwork } from '~~/types/ENetwork'
